refactor workspace: merge into single crate
