+++ .gitignore
@@ -1,3 +1,2 @@
 /target
-/docker/.env
-/docker/.database/*
+.env
