+++ Cargo.toml
@@ -1,4 +1,51 @@
-[workspace]
-members = ["common", "service", "watcher"]
+[package]
+name = "pubky-nexus"
+version = "0.1.0"
+edition = "2021"
+description = "Nexus between homeservers and Pubky-App. Pubky-nexus constructs a social graph out of all of the events on pubky-core homeservers and exposes a social-media-like API capable of powerful Web-of-Trust inference."
+homepage = "https://github.com/pubky"
+repository = "https://github.com/pubky/pubky-nexus"
+license = "MIT"
 
-resolver = "2"
+[dependencies]
+tokio = { version = "1.38.0", features = ["full"] }
+axum = "0.7.5"
+redis = { version = "0.25.4", features = ["tokio-comp"] }
+neo4rs = "0.7.1"
+serde = { version = "1.0.204", features = ["derive"] }
+serde_json = "1.0.120"
+once_cell = "1.19.0"
+utoipa = "4.2.3"
+utoipa-swagger-ui = { version = "7.1.0", features = ["axum"] }
+tower-http = { version = "0.5.2", features = ["fs"] }
+dotenv = "0.15"
+log = "0.4.22"
+env_logger = "0.11.3"
+
+[dev-dependencies]
+anyhow = "1.0.86"
+httpc-test = "0.1.9"
+criterion = { version = "0.5.1", features = ["async_tokio"] }
+
+[lib]
+name = "pubky_nexus"
+path = "src/lib.rs"
+
+[[bin]]
+name = "pubky-nexus-service"
+path = "src/service.rs"
+
+[[bin]]
+name = "pubky-nexus-watcher"
+path = "src/watcher.rs"
+
+[[bench]]
+name = "profile"
+harness = false
+
+# Maximum performance profile
+[profile.release]
+opt-level = 3
+codegen-units = 1
+lto = true
+incremental = false
