--- common/src/db/mod.rs
@@ -1,1 +1,0 @@
-pub mod neo4j;
