--- common/src/lib.rs
@@ -1,4 +1,0 @@
-mod db;
-
-pub mod connectors;
-pub use db::neo4j::queries;
