--- service/.env-sample
@@ -1,21 +1,0 @@
-# Logging
-RUST_LOG=debug
-
-# Service
-SERVER_HOST=127.0.0.1
-SERVER_PORT=8080
-
-# Directory where static files are stored
-STATIC_PATH=./
-
-# Neo4j database
-NEO4J_HOST=localhost
-NEO4J_PORT=7687
-# Neo4J community does not support custom db_name and db_username, the default ones are 'neo4j'
-NEO4J_DB_NAME=neo4j
-NEO4J_DB_USERNAME=neo4j
-NEO4J_PASSWORD=12345678
-
-# Redis database
-REDIS_HOST=localhost
-REDIS_PORT=6379
