--- service/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "pk-social-service"
-version = "0.1.0"
-edition = "2021"
-description = "Server that exposes a social-media-like view out of pubky events for pubky-app clients"
-homepage = "https://github.com/pubky"
-repository = "https://github.com/pubky/pubky-app-backend"
-license = "MIT"
-
-[dependencies]
-pk-social-common = { path = "../common" }
-tokio = { version = "1.38.0", features = ["full"] }
-axum = "0.7.5"
-serde = { version = "1.0.204", features = ["derive"] }
-serde_json = "1.0.120"
-utoipa = "4.2.3"
-utoipa-swagger-ui = { version = "7.1.0", features = ["axum"] }
-tower-http = { version = "0.5.2", features = ["fs"] }
-dotenv = "0.15"
-log = "0.4.22"
-env_logger = "0.11.3"
-
-[dev-dependencies]
-anyhow = "1.0.86"
-httpc-test = "0.1.9"
-criterion = { version = "0.5.1", features = ["async_tokio"] }
-
-[lib]
-name = "pk_social_service"
-path = "src/lib.rs"
-
-[[bin]]
-name = "pk-social-service"
-path = "src/main.rs"
-
-[[bench]]
-name = "profile"
-harness = false
