+++ src/db/mod.rs
@@ -1,2 +1,3 @@
+pub mod connectors;
 pub mod graph;
 pub mod kv;
