+++ src/lib.rs
@@ -1,0 +1,12 @@
+mod config;
+mod db;
+mod error;
+pub mod models;
+pub mod routes;
+mod setup;
+
+pub use config::Config;
+pub use db::graph::queries;
+pub use db::kv::indexes;
+pub use error::{Error, Result};
+pub use setup::setup;
