+++ src/models/profile/counts.rs
@@ -1,13 +1,11 @@
-use pk_social_common::{
-    connectors::{
-        neo4j::get_neo4j_graph,
-        redis::{get_redis_conn, AsyncCommands},
-    },
+use crate::{
+    db::connectors::{neo4j::get_neo4j_graph, redis::get_redis_conn},
     queries,
 };
-
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
+
+use redis::AsyncCommands;
 
 const PROFILE_COUNTS_PREFIX: &str = "profile-counts!";
