+++ src/models/profile/relationship.rs
@@ -1,10 +1,11 @@
-use pk_social_common::connectors::{
-    neo4j::get_neo4j_graph,
-    redis::{get_redis_conn, AsyncCommands},
+use crate::{
+    db::connectors::{neo4j::get_neo4j_graph, redis::get_redis_conn},
+    queries,
 };
-use pk_social_common::queries;
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
+
+use redis::AsyncCommands;
 
 const RELATIONSHIP_PREFIX: &str = "follows!";
