+++ src/routes/static/mod.rs
