--- watcher/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "pk-social-watcher"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
