--- watcher/src/graph/consumer.rs
@@ -1,1 +1,0 @@
-// Reads all the events from the event store
