--- watcher/src/graph/mod.rs
@@ -1,2 +1,0 @@
-pub mod consumer;
-pub mod queries;
