--- watcher/src/kv/queries.rs
@@ -1,1 +1,0 @@
-// Write queries to create the data layer of the app
